Remove psutil dependency
There is no need to check if snapraid in running in this script.
Snapraid already obtains a file based lock, and will exit with a
non-zero exit code when another snapraid process has that lock.

The current code will then raise a SystemError along with the last few
lines of output, which is more than acceptable.